Moved logo and search to their own folder
