+++ frontend/src/components/logo/index.js
@@ -1,0 +1,12 @@
+import React, {Component} from 'react';
+import './style.scss';
+
+class Logo extends Component {
+    render() {
+        return (
+            <div className={"logo"}/>
+        );
+    }
+}
+
+export default Logo;
