+++ frontend/src/components/navigation/search/index.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import './style.scss';
 
-class Search extends Component {
+class Index extends Component {
     constructor(props) {
         super(props);
         this.state = {
@@ -31,4 +31,4 @@ class Search extends Component {
     }
 }
 
-export default Search;
+export default Index;
