+++ frontend/src/components/navigation/search/style.scss
